requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
git-filter-repo==2.45.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.8
importlib==1.0.4
iniconfig==2.0.0
jiter==0.7.0
lxml==5.3.0
numpy==2.1.3
openai==1.53.0
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
PySocks==1.7.1
pytest==8.3.3
requests==2.32.3
rsa==4.9
selenium==4.24.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.66.6
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
vidpy==0.2.1
websocket-client==1.8.0
wsproto==1.2.0
python-dotenv==1.0.0
